+++ src/modules/helpers.js
@@ -1,5 +1,7 @@
+import {popup, btnEditUser, formEdit, formAdd} from './variable.js'
+
 // Кнопка отправки данных юзера
-function renderLoading(isLoading) {
+export function renderLoading(isLoading) {
     if (isLoading) {
       btnEditUser.textContent = "Загрузка...";
     } else {
@@ -8,17 +10,17 @@ function renderLoading(isLoading) {
   }
 
   // Удаление класса с display: none для форм добавления и редактирования профиля
-function formNone() {
+export function formNone() {
     formEdit.classList.remove("popup__form-none");
     formAdd.classList.remove("popup__form-none");
   }
   
   //
-  function popupSwitch() {
+  export function popupSwitch() {
     popup.classList.toggle("popup_is-opened");
   }
   
   // Сброс сообщения ошибки
-  function resError(elem) {
+  export function resError(elem) {
     elem.classList.remove("error-message_active");
   }
